Add integrate_in_range gui functionality
CLI version is working, but some problem with specifing the right
axis in integrate_simpson. So will only work with line-scans currently
GUI version is not doing the operation in place, only printing the
s.data for the integrated spectrum. Need to figure out a way do the
in-place substitution of the spectrum.