ENH: add slider option in navigator argument for plotting
Changed the behavior of s.plot(navigator=None), it used to
make a navigaton window consisting of sliders. Now it does not
create any window for navigation.

Added slider argument, which behaves similarly to the old None
argument.

Currently it will (probably) only work with signals with a signal
dimension of 1.

Will probably need to update some functions which rely on the
sliders to navigate.