BUGFIX: more general dictionary construction for DictionaryTreeBrowser
Constructs dictionaries not only from included DictionaryTreeBrowsers, but from
any items that have 'as_dictionary()' method.
Fixes a bug where , after unfolding a signal and exporting it to dictionary, you have
objects that can be represented as dictionaries but are left as-is.
Because of this bug dictionaries were not passable with pickle.dump and pickle.load
for multiprocessing and similar