BUG-FIX: resume_update was not working as expected.
This was  because `_get_auto_update_plot` was False if suspended. To correct it
`_get_auto_update_plot` was renamed to `_plot_active` and only checks if the
plot is active. `_connect_parameters2update_plot` and `update_plot` now
check if the plot is active, instead of doing it at an upper level.