Made separate component type
Made a separate component type for a gaussian with 'height' instead of
'A'. This makes the parameter vs. peak maximum independent of sigma, and
thereby makes locking of the parameter more viable. As long as there it
no binning, the height parameter corresponds directly to the peak
maximum, if not, the value is scaled by a linear constant
(signal_axis.scale).