Extended ROI use to Model
Replaced to references to _getitem_tuple with a more 'correct' way of
accessing the data. By replacing these two instances, we assure that the
variance Signal actually matches any averaging done on the Signal
itself.

One reference to _getitem_tuple was left as it is, as there is no way to
replace it and still retain the correct functionality. This reference is
in as_signal, where it is being used to extract slices from signal.data
while iterating through the signal.

Also did some cleaning and renaming in DataAxis/AxesManager.