ENH: Rename widgets' "add_axes" to "set_mpl_ax"
Makes it less ambigous what kind of axes that is being used, which will
be especially useful w.r.t. proposed changes for new ROIs, where widgets
also take hyperspy's DataAxis as parameters.

Cons are of course that it will break existing code. Backwards
compatibility with a deprecation warning is possible, and might be
useful.