ENH: rgbx2regular_array keeps masks
`misc.rgb_tools.rgbx2regular_array()` checks if the data is C
contiguous, if not it converts it with  `np.ascontiguousarray()`.
However, if the data is a masked array the mask will dissappear in this
operation. This change allows the data to retain its mask, while still
ensuring C order. It is possible that the entire code could simply be
replaced with `data = data.copy(order='C')` as it is the same for numpy
arrays and masked arrays, however, `np.ascontiguousarray()` might be
more efficient (I don't know).