Fix: component.py's import of traitsui
There where quite a few issues with importing traitsui directly in
component.py. The need for this was because of a bug in traitsui's
handling of a default editor for t.Either(t.CFloat(), Array()). This
change implements a workaround that fixes both problems, but overrides a
function on the trait to do it. As such, it quite a bit of a hack.