Remove sobel filter always constant mode
Setting is really only important at image edges. The default settings
assures that the sobel is always zero at the edges. The previous setting
of `'constant'` is only significant when not using a hanning filter, but
if a hanning is not used it basically introduces a strong edge along the
outside of the image for any non-zero edges.