FIX: make io_plugins/semper_umf work with numpy < v1.9.0
numpy.ndarray.tobytes() was introduced in numpy 1.9.0,
however the current package in Ubuntu 15.10 is numpy 1.8.2.
So changing tobytes() to tostring(), which is in older numpy
versions, should fix this. Without (probably) changing the
functionality.

This will probably also make io_plugins/semper_umf work with
older versions of Ubuntu and Debian.