ENH: Updating plot in `estimate_image_shift`
Previously for Signal.estimate_shift2D(), if the argument plot was True,
a figure would be created for each alignment step. This PR gives the
option to pass a value of 'reuse' to plot, which well then make it reuse
a single plot for each frame. This effectively allows the live
monitoring of the alignment process, but without creating more than one
figure (i.e. only the final frame is accessible after alignment
finishes). This is useful when you have a large alignment job that needs
to be monitored (e.g. to watch for divergence), but where its unfeasible
with hundreds or thousands of figures.

As an alternative, you can also pass a matplotlib.figure.Figure as plot,
which will then be used, but I'm not sure if this has any advantages. It
just simply follows as a "free" alternative because of how the code is
structured.