Fix background removal which was broken for EELS and EDX spectra (#1265)
EELS and EDX signals automatically add some components to the
model on creation. As a result, the automatically created model
for background subtraction was removing a wrong model of the
background. In the case of EELS the error is usually negligible for
core-losses but can be significant for low-losses.

Thank you @AndrewHerzing for reporting this issue.