s.add_marker: remove plot_signal argument. Change plot_marker behaviour
The change in plot_marker behaviour might break some functions.
In the old version, s.add_marker would always call s.plot(), this has
been changed so if plot_marker=False, s.plot() will not be called.
This to allow for permanent markers to be added without opening
a plot.