fix bug in component reversal
Some ICA components remain entirely negative atm. The code seems to look for the maxima and minima in the loadings instead of the factors. With this change it seems to work.
In passing, I also don't know the reason for what seems to me an overly elaborate criterion (line 794). Why not just use:
if maximum< 0:
     self.reverse_bss_component(i)

or even use the mean instead of the maximum, to make the components mostly positive?