Coarse attempt at prefs-based customizable hotkeys
As the field I had to use traits.Directory() in order to
simulate traits.String() or traits.CStr() as they do not
appear to work. Get a KeyError.

Variable names are to be changed. Ctrl + Shift + Arrows
does not work on my home computer. Therefore changed to
Shift + Arrows for now.

Test axes_manager indices changing with hotkey

I did not find a way to test creating a plot and feed simulated
keypresses to it, but feeding equivalent strings to the axes_manager
should be ok.

Added docs, also fixed typos

Coarse example of modifiable hotkeys