Remove advice about the bundle
Since this is no longer the way to get the latest version (if I've understood) I don't think telling people to use the bundle is a good idea.
Incidentally on the "News" page, the links to the bug fixes and enhancements seem to be wrong. They imply that none of either have occurred in this release. I wanted to change that but could not find the file about release 1.3.1.
Maybe I'm jumping the gun and this is all in hand, My apologies if so...