Fix warning when plotting Signal1D with only zero values
When navigating over a multidimensional Signal1D where
some (or all) of the spectra had only zero values, a
warning was raised in the console (or notebook) leading to
the console becoming filled up with warnings.

This was due to the matplotlib subplot ylims being set
to the same values (set_ylim(0, 0)). This solves this
by adding -+ 0.1 to the y_min and y_max, if they're the
same value.

This bug would also occurs if all the values in the spectra
was some other non-zero value.