Fix hanning window calculation
While I very well may be wrong, when taking a 3D array (EELS Spectrum Image) and applying a hanning window for fourier_ratio_deconvolution, the signals have a hanning window applied. This window should be applied to the signal axis, not the navigation axes, but the window size is is calculated using len(hs_object), which is just np.prod(nav_axes_size). This change should (hopefully) calculate the correct hanning window size for the signal axis, such that it matches the documentation. 

If I am wrong, please reject. I also did not know which branch to propose to, and I'm happy to change that if necessary.