Small change, get_current_signal: make it work with LazySignal.__init__
Previously, when self was Lazy, it would call the LazySignal.__init__
signal with `_lazy': False, leading to the non-lazy version of the
signal calling the LazySignal.__init__. However, the non-lazy signal
did not have access to a function which is only available in
LazySignal. This lead to an error, meaning get_current_signal would
fail.